+++ logprep/connector/confluent_kafka/input.py
@@ -37,7 +37,7 @@ from typing import Optional, Tuple, Union
 
 import msgspec
 from attrs import define, field, validators
-from confluent_kafka import (
+from confluent_kafka import (  # type: ignore
     OFFSET_BEGINNING,
     OFFSET_END,
     OFFSET_INVALID,
@@ -47,7 +47,7 @@ from confluent_kafka import (
     Message,
     TopicPartition,
 )
-from confluent_kafka.admin import AdminClient
+from confluent_kafka.admin import AdminClient  # type: ignore
 
 from logprep.abc.connector import Connector
 from logprep.abc.input import (
@@ -318,19 +318,19 @@ class ConfluentKafkaInput(Input):
         self.metrics.number_of_errors += 1
         logger.error(f"{self.describe()}: {error}")
 
-    def _stats_callback(self, stats: str) -> None:
+    def _stats_callback(self, stats_raw: str) -> None:
         """Callback for statistics data. This callback is triggered by poll()
         or flush every `statistics.interval.ms` (needs to be configured separately)
 
         Parameters
         ----------
-        stats : str
+        stats_raw : str
             statistics from the underlying librdkafka library
             details about the data can be found here:
             https://github.com/confluentinc/librdkafka/blob/master/STATISTICS.md
         """
 
-        stats = self._decoder.decode(stats)
+        stats = self._decoder.decode(stats_raw)
         self.metrics.librdkafka_age += stats.get("age", DEFAULT_RETURN)
         self.metrics.librdkafka_rx += stats.get("rx", DEFAULT_RETURN)
         self.metrics.librdkafka_tx += stats.get("tx", DEFAULT_RETURN)
@@ -395,7 +395,7 @@ class ConfluentKafkaInput(Input):
         base_description = super().describe()
         return f"{base_description} - Kafka Input: {self._config.kafka_config['bootstrap.servers']}"
 
-    def _get_raw_event(self, timeout: float) -> bytearray:
+    def _get_raw_event(self, timeout: float) -> bytes | None:
         """Get next raw Message from Kafka.
 
         Parameters
@@ -429,7 +429,7 @@ class ConfluentKafkaInput(Input):
         self.metrics.current_offsets.add_with_labels(message.offset() + 1, labels)
         return message.value()
 
-    def _get_event(self, timeout: float) -> Union[Tuple[None, None], Tuple[dict, dict]]:
+    def _get_event(self, timeout: float) -> Tuple[None, None] | Tuple[dict, bytes]:
         """Parse the raw document from Kafka into a json.
 
         Parameters