+++ tests/acceptance/test_full_configuration.py
@@ -16,7 +16,6 @@ from tests.acceptance.util import (
     get_full_pipeline,
     start_logprep,
     stop_logprep,
-    wait_for_output,
 )
 
 
@@ -35,7 +34,7 @@ def test_start_of_logprep_with_full_configuration_from_file(tmp_path):
     while True:
         assert not re.search("Invalid", output)
         assert not re.search("Exception", output)
-        assert not re.search("critical", output)
+        assert not re.search("Critical", output)
         assert not re.search("Error", output)
         assert not re.search("ERROR", output)
         if re.search("Startup complete", output):
@@ -58,7 +57,7 @@ def test_start_of_logprep_with_full_configuration_http():
         while True:
             assert not re.search("Invalid", output), output
             assert not re.search("Exception", output), output
-            assert not re.search("critical", output), output
+            assert not re.search("Critical", output), output
             assert not re.search("Error", output), output
             assert not re.search("ERROR", output), output
             if re.search("Startup complete", output):
@@ -105,8 +104,16 @@ output:
     }
     with HTTPServerForTesting.run_in_thread():
         proc = start_logprep("${LOGPREP_API_ENDPOINT}/generated_config.yml", env=env)
-        wait_for_output(proc, "Startup complete")
-        stop_logprep(proc)
+        output = proc.stdout.readline().decode("utf8")
+        while True:
+            assert not re.search("Invalid", output)
+            assert not re.search("Exception", output)
+            assert not re.search("Critical", output)
+            assert not re.search("Error", output)
+            assert not re.search("ERROR", output)
+            if re.search("Startup complete", output):
+                break
+            output = proc.stdout.readline().decode("utf8")
 
 
 def test_logprep_exposes_prometheus_metrics_and_healthchecks(tmp_path):
@@ -157,7 +164,7 @@ def test_logprep_exposes_prometheus_metrics_and_healthchecks(tmp_path):
     while True:
         output = proc.stdout.readline().decode("utf8")
         assert "error" not in output.lower(), "error message"
-        assert "critical" not in output.lower(), "error message"
+        assert "Critical" not in output.lower(), "error message"
         assert "exception" not in output.lower(), "error message"
         assert not re.search("Shutting down", output)
         if "Startup complete" in output: